requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
pandas==2.2.3
numpy==1.24.4
fastparquet==2024.11.0
scikit-learn==1.6.1
scikit-surprise==1.1.4
